requirements.txt
langchain 
wikipedia 
langchain-community 
langchain-openai
langchain-anthropic
python-dotenv
pydantic
duckduckgo-search
google-api-python-client
google_auth_oauthlib
google-auth-httplib2